Relocate system jobs to sub-files
